redid mate with qtree, fixed world life
didn't see much of a performance increase, i think the lag has more to do with clusters of critters getting stuck together and checking against each other. should test with bigger world // more separation. but for now it's fine, i don't think this will be an issue in the bigger world