db fuckery, compaction
critter threshold back to around 600.

i think the separation of dbs is prob not ideal, unless i update both eco and critter dbs often there's a chance for mismatch, but if i do that, what was the point of separating? keeping them separate for now, rn the crash edge case doesn't outweigh the stuff i need to have done for the playtest

nvm reverting critter db stuff because it was the FUCKING ANCESTRY stuff that was so big. also updating eco db every min just and at donation/baby because im tired of tweaking. had to add manual compaction because of db file size if runs too long. also added console backup option.

~~frankensteining critters from two dbs now, methods for static or dynamic critter info.~~